refactor: input,output controller 합치기
