--- src/main/java/lotto/controller/LottoOutputController.java
@@ -1,51 +1,0 @@
-package lotto.controller;
-
-import lotto.dto.LottoGameInfo;
-import lotto.domain.lotto.LottoRank;
-import lotto.domain.lotto.Lottos;
-import lotto.dto.LottoResult;
-import lotto.view.output.LottoBuyOutputView;
-import lotto.view.output.LottoResultOutputView;
-
-import java.util.Map;
-
-public class LottoOutputController {
-    private final LottoBuyOutputView lottoBuyOutputView;
-    private final LottoResultOutputView lottoResultOutputView;
-
-    public LottoOutputController() {
-        this.lottoBuyOutputView = new LottoBuyOutputView();
-        this.lottoResultOutputView = new LottoResultOutputView();
-    }
-
-    public void printBuyLottos(LottoGameInfo lottoGameInfo) {
-        Lottos lottos = lottoGameInfo.getLottos();
-        lottoBuyOutputView.printPurchaseLottoCountMessage(lottos);
-        lottoBuyOutputView.printAllLottoNumbers(lottos);
-    }
-
-    public void resultMessage(LottoGameInfo lottoGameInfo, LottoResult lottoResult) {
-        Map<LottoRank, Long> lottoRankCountMap =
-                printResult(lottoGameInfo.getLottos(), lottoResult);
-
-        printTotalProfitRate(lottoGameInfo.getCash(), lottoRankCountMap);
-    }
-
-
-    private Map<LottoRank, Long> printResult(Lottos lottos, LottoResult lottoResult) {
-        Map<LottoRank, Long> lottoRankCountMap = lottos.getLottoRankCountMap(lottoResult);
-        printStartResultMessage();
-        lottoResultOutputView.printLottoAllMatch(lottoRankCountMap);
-
-        return lottoRankCountMap;
-    }
-
-    private void printStartResultMessage() {
-        lottoResultOutputView.printWinningStatisticsStartMessage();
-        lottoResultOutputView.printDashLine();
-    }
-
-    private void printTotalProfitRate(long amount, Map<LottoRank, Long> lottoRankCountMap) {
-        lottoResultOutputView.printTotalProfitRate(amount, lottoRankCountMap);
-    }
-}
